Install now has textbox's for Website Name/URL, Cookie Domain/Path, and uses the Admin email provided for the contact link.
All are addslash'ed for security and enclosed in a collapsable/expandable box with the appropriate JScript in general.php and included in the script via resources/output.php.

Images (currently 5) are now in images/ directory, and various little changes have been made to the wording of the install.

Still left to do is make the box's collapsed by default (tried but no success), and add some checking, to make sure the required feilds are filled in.