Converters - attachment types module
Don't have IPB2 downloaded yet, so that will come later.  I haven't found the setting for XMB, so if I find it it will come later.  Attachments doesn't come with phpBB2, so naturally there will be no attachment types.  Same with punBB.

I needed to modify some of the other files in order to get the converter to work so that I can test them.  Please revert if someone was in the middle of changes.

Tikitiki you owe me a Christmas and New Years present.

Merry Christmas and Happy New Year to all staff.

This is a 1am production, so there may be a small amount of Dennis features. :)