I thought the pm popup issue was only with the new popup, not the old, as it says in the report.
I only fixed the popupmenu the way we do it on the main forum. And it looks stupid if a user doesn't have javascript enabled anyways. The dropdown button should not show if the user doesn't have javascript enabled. Adding style=\"display: none;\" to that then - and showing it again in the same as statement as stop it from showing the settings when javascript enabled.

BTW chris, you do realise javascript can't change the display attribute if it gets added by CSS extra, not html style css? You have to add it to the element. Now when you click the button, nothing happens. :)