Bug Fix: http://beta.mybboard.net/mybb_140/thread-1064.html "the export didn't export the CSS files"
Included suggestion to disable logging of emails because in some countries its illegal.
Included suggestion to better handle RTL in quotes
Included more links on theme page to edit pages
Included JS effect to remove the "inherited warning" when it's saved in the simple editor
Turning off "Delayed Thread Views" task by default as it is handled by the settings.