First round of jQuery updates
MyBB 1.8 replaces Prototype with jQuery. This first round of updates adds jQuery itself, removes Prototype and Prototype related files,
and replaces the old autocomplete with Typeahead.js. This commit is also a start for purging MyBB of XML HTTP requests. We've opted for JSON instead
as it works nicely with newer technologies.