Begin rewriting general.js and captcha.js
As a part of the jQuery conversion, this commit provides an uncomplete version of general.js. It still requires some work (standards, performance) but is generally usable. It lacks MyBB.popupWindow() which is being researched, and all functions which use it. Lang is also planned to be rewritten but remains unchanged until then. The activity indicator will be added in a later commit.

Additionally, captcha.js has been rewritten and works.