Make MyBB::dev_mode useful
MyBB has a dev_mode class variable which, when true, fetches the templates directly from mybb_theme.xml. This is great, but you have to set it yourself to use it. This commit checks for an environment variable (MYBB_DEV_MODE) and sets dev_mode to true if it exists.

Additionally, if dev_mode is true allow the debug information to be shown to all usergroups including guests. This should help improve debugging.