Update "5 Latest Moderator Actions" no results message
In MyBB 1.6, we show "Sorry, there were no results found with the criteria you selected." in the Mod CP when there aren't any moderator logs in the database. This makes it sound like the moderator has performed an action when they haven't. To make it sound more friendly and understandable, I've added a new language string and template which displays "No moderator actions are currently logged." instead.