Use `text-overflow: ellipsis` for session IP address
Due to the recent addition of IPv6 support the IP addresses of the administrators logged in will jump to a new line if they're using IPv6. To make this a little neater I've added `text-overflow: ellipsis` to the desktop and mobile classes, and given the <li> a title with the full IP address.