Improve user banning
Currently when banning users, you're presented with an input field which allows you to type as much as you like. The hard limit here is 225 characters. When writing or editing a ban reason you can't always see the whole message, which is a bit silly. Switch to a textarea for ban reasons and enforce a maximum length with the HTML maxlength attribute and server side with my_substr(). We also use my_wordwrap() on displayed ban reasons so we don't stretch the page out.

Additionally, this adds a maxlength option to the DefaultForm::generate_text_area() method.