Make thread notes usable
In mid-2012 there was an internal discussion about how unusable thread notes were and how little they were used because of this. In an aim to improve this, and making moderation in general easier, I've added a link to the thread notes at the top of the thread for moderators. This makes it clear when there are any thread notes but is still unobtrusive. The behavior of thread notes has changed and you can now directly access the page with a GET request to `moderation.php?action=threadnotes&modtype=thread&tid={tid}`. Additionally the Thread Modes have been modified and are now accessible using a popup menu. As Threaded Mode is unused by most and will be removed in a future release it makes sense to tuck these away.

Introduces 1 new language string and 1 new template.