Execute queries on the master database if the last write query was to the master
This is a port of the work by @ImadJomaa which never reached the 1.6 series.

Internal thread: http://community.mybb.com/thread-119212.html

As I understand it, there were a few incorrect comments about which database was being written to (master/slave) and the query() method would always query the slave. If the last query was a write query to the master, the next query should be to the master too due to latency with slaves being updated.