Improve collapsed style
Currently if a user was to collapse a table with a thead, the bottom
corners have no boarder radius, but the container does leaving it
looking out of place.
This commit adds the thead_collapsed class to the .thead (if present)
when collapsed or when set as collapsed from a cookie.