Working on #251 jQuery Conversion
Fixed PM Tickbox not working properly. Actually almost fixed it. Because
there's still a bug which won't display the message in the textarea
unless the box is shown by default (which isn't). This happens because
the SCEditor textarea height is 0px and therefore we can't see the
message! If we refresh with the box ticked, the height is no longer 0px
and we can see the message.