Fix #1453
Used addslashes() to prevent JS errors (not sure if it's the best
solution, but it was used inmisc.php only for some reason).

Also added spaces in misc.php inserter for consistency and listed the
replacements in new lines instead of sepearating them with space/comma
which looks kinda weird.