Fix #1373 profile field permissions
Adds missing `viewableby`/`editableby`/`postnum` checks.

Hides empty fields on postbit and fields that shouldn't be displayed on
profile.

`viewableby`/`editableby` set to All for new fields by default (not sure
why they were None in the first place).

Fixes regex error when a field is not required and not filled.

https://github.com/mybb/mybb/issues/1373