Fix #2003 Deleting an event in 'editevent' without checking checkbox throws errors
Adds a new action 'do_deleteevent' for deleting events and displays a proper error message (with a new language string) when it is not checked. The JavaScript function MyBB.deleteEvent(eid) is also using this action now. The action 'do_editevent' is not capable of deleting events anymore.

https://github.com/mybb/mybb/issues/2003

Bumps version of general.js and sets version to 1806 on the modified template

Updates all links to general.js to version 1806.

Bumps headerinclude version to 1806.

Adds plugin hooks `calendar_do_deleteevent_start` and `calendar_do_deleteevent_end`