Change protocol-relative embeds to HTTPS default
Changes jscripts/bbcodes_sceditor.js to use https URLs instead of
protocol-relative URLs. There is no technical reason to not use
secure versions of resources, as unless one is in the middle of a
man-in-the-middle attack, their request is sent much more securely.

If the user is in a MITM scenario, they're already screwed with
the protocol-relative URLs. That's not the point of this change.
They're screwed even with this change -- I get that.

Most users are not under active MITM attacks, and most forums are not
running a TLS configuration. As a result, users' resource requests
effectively get downgraded to HTTP for no reason, exposing request
metadata that may pose privacy risks.

Signed-off-by: Josh Harmon <me@joshharmon.me>