Moderators can edit soft deleted posts.
Moderators should be able to edit soft deleted posts again as long as they also have permission to view deleted posts. It's a bit silly a mod can view the a soft deleted post but they cannot edit it. There are many instances in which modifying a soft deleted post will come in handy.