Fixes issues with board language setting being overwritten
`$mybb->settings['bblanguage']` is used to hold the current language being used by the user, the original default language is held in `$mybb->settings['orig_bblanguage']`

This commit also reduces the repeated logic for setting `$mybb->settings['bblanguage']` at the beginning of requests when the session and language are setup. Made global.php, archive/global.php and xmlhttp.php consistent for setting up the language and set `$mybb->session` to reference `$session` where this was not done before.