Fixes #3618: Attachments not working properly in latest version (#3986)
The main problem was that in some (all?) cases, PHP upload
errors were being ignored. This was because the check for, and parsing
of, PHP upload errors occurred in the function upload_attachment()
which was sometimes not called on PHP upload errors due to the
following conditional in add_attachments():

if(!empty($FILE['name']) && !empty($FILE['type']))

More specifically, $FILE['type'] can be set to empty even though a
file with a valid file type was uploaded which exceeded the php.ini
setting upload_max_filesize, thus generating an UPLOAD_ERR_INI_SIZE
error.

This refactoring of the code ensures that the check for, and parsing of,
PHP upload errors is moved into a separate function, called up-front,
before that conditional, so that they cannot be ignored.