Fix #4184 and (partially) #4182 - SCEditor-related bugs
Fix #4184 - redundant tags on paste of unformatted text - by ignoring
fonts, colors, and sizes that match the editor's default.

Partially fix (client-side only) #4182 - double quotes in font lists -
by replacing double quotes in fonts with single quotes.

Commit both fixes at once because doing so separately would be likely to
cause a confusing conflict that would then need to be resolved (due to
the change to the condition in the 'font' method required by the fix for