Rework #4342 - Relative/absolute path (#4370)
[Rebased for 1.9 by Laird]

Introduce a `mk_path_abs()` function so that the `uploadspath` setting
can be "absolutised" on-demand rather than at the beginning of
processing.

Resolves the first item of #4360 without the need for manipulating the
setting's value in the proposed fix in #4361. If merged, provides the
basis for resolving the conversation[1] in that PR so that it can be
adapted a little and continue to resolve the other items of #4360.

[1] https://github.com/mybb/mybb/pull/4361/files#r617217364