Improve the logic and caching of resolve_themelet_resource()
* Logic: all plugins were being checked for the resource, instead of
just the plugin that owns the resource.

* Caching:

  * Leverage the existing `getIncludedFiles()` method of the SCSS
compilation result, storing it in a dependencies (.deps) file in the
cache directory, rather than running the custom function I'd written to
generate those same results
on every page load.

* Also store a dependencies file in the cache directory for every
resource, so that we know the cache is stale in the case that the target
resolved resource changes - e.g., if a new board theme adds an override
of that resource. The first line of the dependency file is the primary
resource. Additional lines are dependencies, such as for SCSS `@import`
files.

* Additionally: refactor the code a little to avoid redundancy, by
abstracting it into the little helper function `test_set_path()`.