Mostly remove support for DB themes/templates
Add preliminary support for filesystem-based themes only. Some
functionality still needs to be migrated, with the odd error occurring
on some pages, especially in the ACP.

The essential changes are in `global.php`.

Also of note is that the install/upgrade routines now skip/drop the
theme/template-related DB tables.

The bulk of this commit is simply removing assignments/modifications to
`$templatelist`.