Call the provided `twig_include` extension from our override
For some reason, this doesn't seem to be necessary anyway, because
despite my confirming that the provided function doesn't run unless we
explicitly call it here, functionality did not seem to be in any way
affected, which is quite perplexing. It still seems to be proper and
good practice to add this call anyway.