Improve conditional matching from `resources.json`
The template context(s) in which the conditional matches are tested are
now specific to that template rather than to the highest level template.
That is, conditional matching now supports the context(s) of a template
that was included by another template.

Note that this functionality is not currently used, nor expected to be
used much if at all, but it adds flexibility, so it seems worth keeping.